greatly improve accuracy of the caclualted average pixel coordinate by shrinking the bounding box of the legend

+++ ColorCodedLegendUtil/JSON/LegendBoxEntry.cs
@@ -8,5 +8,5 @@ public class LegendBoxEntry
     public string FileName { get; set; } = default!;
 
     [JsonPropertyName("legend_bbox")]
-    public float[] LegendBBox { get; set; } = new float[4]; // [X1, Y1, X2, Y2]
+    public int[] LegendBBox { get; set; } = new int[4]; // [X1, Y1, X2, Y2]
 }
